Download submissions as JSON
